moved emitted_event to main folder of types
