--- crates/types/src/models/mod.rs
@@ -1,1 +1,0 @@
-pub mod emitted_event;
